Add api to get thermal data
